+++ app/scripts/controllers/bridge/constants.ts
@@ -1,0 +1,7 @@
+import { BridgeControllerState, BridgeFeatureFlagsKey } from './types';
+
+export const DEFAULT_BRIDGE_CONTROLLER_STATE: BridgeControllerState = {
+  bridgeFeatureFlags: {
+    [BridgeFeatureFlagsKey.EXTENSION_SUPPORT]: false,
+  },
+};
