+++ app/scripts/controllers/bridge/types.ts
@@ -1,0 +1,16 @@
+// Maps to BridgeController function names
+export enum BridgeBackgroundAction {
+  SET_FEATURE_FLAGS = 'setBridgeFeatureFlags',
+}
+
+export enum BridgeFeatureFlagsKey {
+  EXTENSION_SUPPORT = 'extensionSupport',
+}
+
+export type BridgeFeatureFlags = {
+  [BridgeFeatureFlagsKey.EXTENSION_SUPPORT]: boolean;
+};
+
+export type BridgeControllerState = {
+  bridgeFeatureFlags: BridgeFeatureFlags;
+};
